requirements.txt
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
Babel==2.10.3
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
certvalidator==0.11.1
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.5
decorator==5.1.1
docopt==0.6.2
docutils==0.20.1
ebaysdk==2.1.5
endesive==2.17.2
executing==2.1.0
Flask==3.1.0
freezegun==1.2.1
future==1.0.0
geoip2==2.9.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
ipykernel==6.29.5
ipython==8.27.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.2
jupyter_core==5.7.2
libsass==0.22.0
lxml==5.2.1
lxml_html_clean==0.4.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
maxminddb==2.6.2
nest-asyncio==1.6.0
num2words==0.5.13
numpy==2.1.0
ofxparse==0.21
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plum-dispatch==1.7.4
polib==1.1.1
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydot==1.4.2
Pygments==2.18.0
PyKCS11==1.5.17
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.2.0
PyPDF2==2.12.1
pypng==0.20220715.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-ldap==3.4.4
python-stdnum==1.19
pytz==2022.1
pyusb==1.2.1
pyzmq==26.2.0
qrcode==7.4.2
reportlab==4.1.0
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.2.0
setuptools==74.1.2
six==1.16.0
soupsieve==2.6
Spire.Pdf==10.8.1
stack-data==0.6.3
tkintertable==1.3.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.32.1
vobject==0.9.6.1
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.44.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
zeep==4.2.1
zope.event==5.0
zope.interface==7.1.1
